Part 4: Exercise 4.21

+++ part4/controllers/blogs.js
@@ -45,6 +45,24 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+	if (!decodedToken) {
+		return response.status(400).json({ error: 'token invalid' })
+	}
+
+	const blog = await Blog.findById(request.params.id)
+
+	if (!blog) {
+		return response.status(404).json({ error: 'blog not found' })
+	}
+
+	if (blog.user.toString() !== decodedToken.id.toString()) {
+		return response
+			.status(400)
+			.json({ error: 'only the creator can delete this blog' })
+	}
+
 	await Blog.findByIdAndDelete(request.params.id)
 	response.status(204).end()
 })
